// src/common/types.ts
export type NonEmptyArray<T> = [T, ...T[]];
// ...
export type CompositeWeights<RatingKey = string> = {
	[key: string]: {
		ratings: (RatingKey | number)[];
		weights?: number[];
		skill?: {
			label: string;
			cutoff?: number;
		};
	};
};
// ...
export type Conf = { cid: number; name: string };
export type Div = { cid: number; did: number; name: string };